+++ vistasFacebook/src/main/java/comVista/IComunicadorVista.java
@@ -18,7 +18,7 @@ public interface IComunicadorVista {
     public void EditarUsuario(Usuario usuario);
     public void registrarPublicacion(Publicacion publicacion);
     public void consultarPublicaciones();
-    public void cosultarUsuarioPorId(Integer id);
+    public void cosultarUsuarioPorId(Long id);
     public void eliminarUsuario(Usuario usuario);
     public void eliminarPublicacion(Publicacion publicacion);
     public void eliminarComentario(Comentario comentario);
